requirements.txt
appdirs==1.4.3
coreapi==2.3.0
coreschema==0.0.4
Django==1.10.6
django-filter==1.0.1
django-rest-auth==0.9.1
django-rest-swagger==2.1.2
djangorestframework==3.6.2
itypes==1.1.0
Jinja2==2.9.5
Markdown==2.6.8
MarkupSafe==1.0
openapi-codec==1.3.1
packaging==16.8
pkg-resources==0.0.0
pyparsing==2.2.0
requests==2.13.0
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
psycopg2==2.6.2
invoke==0.13.0